+++ tests/e2e/transaction/transactionSubmission.test.ts
@@ -7,7 +7,7 @@ import {
   Deserializer,
   SigningSchemeInput,
   MultiKey,
-  AccountAuthenticatorMultiKey,
+  MultiKeyAccount,
   RawTransaction,
   TransactionPayloadEntryFunction,
   Bool,
@@ -603,37 +603,59 @@ describe("transaction submission", () => {
         signaturesRequired: 2,
       });
 
-      const authKey = multiKey.authKey();
+      const account = new MultiKeyAccount({
+        multiKey,
+        signers: [singleSignerED25519SenderAccount, singleSignerSecp256k1Account],
+      });
 
-      const multiKeyAccountAddress = authKey.derivedAddress();
-
-      await aptos.fundAccount({ accountAddress: multiKeyAccountAddress, amount: 100_000_000 });
+      await aptos.fundAccount({ accountAddress: account.accountAddress, amount: 100_000_000 });
 
       const transaction = await aptos.transaction.build.simple({
-        sender: multiKeyAccountAddress,
+        sender: account.accountAddress,
         data: {
           function: `0x${contractPublisherAccount.accountAddress.toStringWithoutPrefix()}::transfer::transfer`,
           functionArguments: [1, receiverAccounts[0].accountAddress],
         },
       });
-      // create a bitmap where singleSignerED25519SenderAccount and singleSignerSecp256k1Account
-      const bitmap = multiKey.createBitmap({ bits: [0, 2] });
 
-      // account1 and account3 sign the transaction
-      const account1Authenticator = aptos.transaction.sign({ signer: singleSignerED25519SenderAccount, transaction });
-      const account3Authenticator = aptos.transaction.sign({ signer: singleSignerSecp256k1Account, transaction });
+      const senderAuthenticator = aptos.transaction.sign({ signer: account, transaction });
 
-      if (!(account1Authenticator.isSingleKey() && account3Authenticator.isSingleKey())) {
-        throw new Error("Both AccountAuthenticators should be an instance of AccountAuthenticatorSingleKey");
-      }
+      const response = await aptos.transaction.submit.simple({ transaction, senderAuthenticator });
+      await aptos.waitForTransaction({
+        transactionHash: response.hash,
+      });
+      expect(response.signature?.type).toBe("single_sender");
+    });
 
-      const multiKeyAuth = new AccountAuthenticatorMultiKey(
+    test("it submits a multi key transaction with misordered signers", async () => {
+      const multiKey = new MultiKey({
+        publicKeys: [
+          singleSignerED25519SenderAccount.publicKey,
+          legacyED25519SenderAccount.publicKey,
+          singleSignerSecp256k1Account.publicKey,
+        ],
+        signaturesRequired: 2,
+      });
+
+      const account = new MultiKeyAccount({
         multiKey,
-        [account1Authenticator.signature, account3Authenticator.signature],
-        bitmap,
-      );
+        // the input to signers does not maintain ordering
+        signers: [singleSignerSecp256k1Account, singleSignerED25519SenderAccount],
+      });
 
-      const response = await aptos.transaction.submit.simple({ transaction, senderAuthenticator: multiKeyAuth });
+      await aptos.fundAccount({ accountAddress: account.accountAddress, amount: 100_000_000 });
+
+      const transaction = await aptos.transaction.build.simple({
+        sender: account.accountAddress,
+        data: {
+          function: `0x${contractPublisherAccount.accountAddress.toStringWithoutPrefix()}::transfer::transfer`,
+          functionArguments: [1, receiverAccounts[0].accountAddress],
+        },
+      });
+
+      const senderAuthenticator = aptos.transaction.sign({ signer: account, transaction });
+
+      const response = await aptos.transaction.submit.simple({ transaction, senderAuthenticator });
       await aptos.waitForTransaction({
         transactionHash: response.hash,
       });