+++ tests/e2e/transactionManagement/transactionWorker.test.ts
@@ -1,5 +1,5 @@
 import { longTestTimeout } from "../../unit/helper";
-import { Account } from "../../../src/core";
+import { Account } from "../../../src/account";
 import { InputGenerateTransactionPayloadData } from "../../../src/transactions/types";
 import { TransactionWorker } from "../../../src/transactions/management/transactionWorker";
 import { TransactionResponseType, TypeTagAddress, TypeTagU64 } from "../../../src";